# API/docs/swagger.yaml
basePath: /
definitions:
  controllers.AgendaRequest:
    properties:
      convenio_id:
        example: 1
        type: integer
      data_hora:
        example: "2024-01-15T10:00:00Z"
        type: string
      observacoes:
        example: Paciente com alergia
        type: string
      paciente_id:
        example: 1
        type: integer
      procedimento_id:
        example: 1
        type: integer
      usuario_id:
        example: 1
        type: integer
    required:
    - data_hora
    - paciente_id
    - procedimento_id
    - usuario_id
    type: object
  controllers.AgendaResponse:
    properties:
      clinica_id:
        example: 1
        type: integer
      convenio_id:
        example: 1
        type: integer
      data_hora:
        example: "2024-01-15T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      observacoes:
        example: Paciente com alergia
        type: string
      paciente_id:
        example: 1
        type: integer
      procedimento_id:
        example: 1
        type: integer
      status_id:
        example: 1
        type: integer
      usuario_id:
        example: 1
        type: integer
    type: object
...
  controllers.AtualizarStatusRequest:
    properties:
      status_id:
        example: 2
        type: integer
    required:
    - status_id
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        example: Erro na operação
        type: string
    type: object
...
  controllers.MessageResponse:
    properties:
      message:
        example: Operação realizada com sucesso
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API multiclínica com isolamento por tenant (clinic_id), RBAC por papel
    (ADM_GERAL, DONO, MEDICO, SECRETARIA), agendamento com validação de conflitos,
    prontuário com imutabilidade após 24h, auditoria LGPD e métricas de faturamento
    e no-show.
  title: API White Label — Clínicas
  version: "1.0"
paths:
  /admin/assinaturas:
    get:
      description: Lista todas as assinaturas
      parameters:
      - description: Filtrar por ativo
        in: query
        name: ativo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar assinaturas
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Cria uma nova assinatura
      parameters:
      - description: Dados da assinatura
        in: body
        name: assinatura
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Criar assinatura
      tags:
      - Admin
...
  /clinicas/agenda:
    get:
      consumes:
      - application/json
      description: Lista todos os agendamentos da clínica
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.AgendaResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar agendamentos
      tags:
      - Agenda
    post:
      consumes:
      - application/json
      description: Cria um novo agendamento na clínica
      parameters:
      - description: Dados do agendamento
        in: body
        name: agendamento
        required: true
        schema:
          $ref: '#/definitions/controllers.AgendaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.AgendaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar agendamento
      tags:
      - Agenda
  /clinicas/agenda/{id}/status:
    put:
      consumes:
      - application/json
      description: Atualiza o status de um agendamento
      parameters:
      - description: ID do agendamento
        in: path
        name: id
        required: true
        type: integer
      - description: Novo status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controllers.AtualizarStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar status do agendamento
      tags:
      - Agenda
  /clinicas/agenda/horarios-disponiveis:
    get:
      consumes:
      - application/json
      description: Buscar horários disponíveis para agendamento
      parameters:
      - description: ID do usuário
        in: query
        name: usuario_id
        required: true
        type: integer
      - description: ID do procedimento
        in: query
        name: procedimento_id
        required: true
        type: integer
      - description: Data no formato YYYY-MM-DD
        in: query
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Horários Disponíveis
      tags:
      - Agenda
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"